refactor(App-UserDelete): change name of file and location
